+++ src/main/java/nexusvault/cli/extensions/convert/ConverterExtension.java
@@ -191,7 +191,12 @@ public final class ConverterExtension extends AbstractExtension {
 		final var factories = createConverterFactory(factoryIds);
 		applyConvertersArgs(factories, options);
 		final var converters = createConverters(factories);
-		return convert(requests, converters, callback);
+		try {
+			final var results = convert(requests, converters, callback);
+			return results;
+		} finally {
+			converters.forEach((k, v) -> v.deinitialize());
+		}
 	}
 
 	private Path makeOutputDir(final Path rootOutputDir, final ConversionRequest request) {
@@ -204,8 +209,8 @@ public final class ConverterExtension extends AbstractExtension {
 			} else {
 				outputDir = rootOutputDir.resolve(outputDir);
 			}
-
-			final var inputDir = request.input.getDirectory().resolve(PathUtil.getFileName(request.input.getFile()));
+      
+			final var inputDir = request.input.getFilePath().resolveSibling(PathUtil.getFileName(request.input.getFile()));
 			if (!inputDir.isAbsolute()) {
 				outputDir = outputDir.resolve(inputDir);
 			} else {