+++ src/main/java/nexusvault/cli/extensions/convert/IsArgument.java
@@ -20,4 +20,5 @@ public @interface IsArgument {
 
 	String name();
 
+	boolean isArgumentOptional() default false;
 }