+++ src/main/java/nexusvault/cli/extensions/convert/converter/m3/M32CsvFactory.java
@@ -1,12 +1,10 @@
 package nexusvault.cli.extensions.convert.converter.m3;
 
-import nexusvault.cli.core.AutoInstantiate;
 import nexusvault.cli.extensions.convert.Converter;
 import nexusvault.cli.extensions.convert.ConverterArgs;
 import nexusvault.cli.extensions.convert.ConverterFactory;
 import nexusvault.cli.extensions.convert.IsFactory;
 
-@AutoInstantiate
 @IsFactory(id = "m32csv", fileExtensions = "m3")
 public final class M32CsvFactory implements ConverterFactory {