+++ src/main/java/nexusvault/cli/extensions/convert/converter/m3/M32GltfFactory.java
@@ -1,6 +1,5 @@
 package nexusvault.cli.extensions.convert.converter.m3;
 
-import nexusvault.cli.core.AutoInstantiate;
 import nexusvault.cli.core.cmd.ArgumentHelper;
 import nexusvault.cli.extensions.convert.Converter;
 import nexusvault.cli.extensions.convert.ConverterArgs;
@@ -8,13 +7,12 @@ import nexusvault.cli.extensions.convert.ConverterFactory;
 import nexusvault.cli.extensions.convert.IsArgument;
 import nexusvault.cli.extensions.convert.IsFactory;
 
-@AutoInstantiate
 @IsFactory(id = "m32gltf", fileExtensions = "m3", priority = 3)
 public final class M32GltfFactory implements ConverterFactory {
 
 	private boolean includeTextures = true;
 
-	@IsArgument(name = "m32gltf-textures")
+	@IsArgument(name = "m32gltf-textures", isArgumentOptional = true)
 	public void setIncludeTextures(boolean value) {
 		this.includeTextures = value;
 	}
@@ -30,10 +28,6 @@ public final class M32GltfFactory implements ConverterFactory {
 
 	@Override
 	public void applyArguments(ConverterArgs args) {
-		if (args.has("m32gltf-textures")) {
-			setIncludeTextures(ArgumentHelper.toBoolean("m32gltf-textures", getIncludeTextures()));
-		}
-
+		args.onHas("m32gltf-textures", value -> setIncludeTextures(ArgumentHelper.toBoolean(value, getIncludeTextures())));
 	}
-
 }