+++ src/main/java/nexusvault/cli/extensions/convert/converter/tbl/Csv2TblFactory.java
@@ -1,15 +1,12 @@
 package nexusvault.cli.extensions.convert.converter.tbl;
 
-import nexusvault.cli.core.AutoInstantiate;
 import nexusvault.cli.extensions.convert.Converter;
 import nexusvault.cli.extensions.convert.ConverterArgs;
 import nexusvault.cli.extensions.convert.ConverterFactory;
 import nexusvault.cli.extensions.convert.IsArgument;
-import nexusvault.cli.extensions.convert.IsFactory;
 import nexusvault.format.tbl.converter.CSV;
 
-@AutoInstantiate
-@IsFactory(id = "csv2tbl", fileExtensions = "csv")
+// @IsFactory(id = "csv2tbl", fileExtensions = "csv")
 public final class Csv2TblFactory implements ConverterFactory {
 
 	private String cellDelimiter = ";";
@@ -25,7 +22,7 @@ public final class Csv2TblFactory implements ConverterFactory {
 
 	@Override
 	public void applyArguments(ConverterArgs args) {
-		setCellDelimiter(args.getOrElse("csv2tbl-delimiter", getCellDelimiter()));
+		args.onHas("csv2tbl-delimiter", value -> setCellDelimiter(value));
 	}
 
 	@Override