+++ src/main/java/nexusvault/cli/extensions/convert/converter/tex/Tex2Png.java
@@ -5,6 +5,7 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
+import java.util.LinkedList;
 
 import javax.imageio.ImageIO;
 
@@ -18,14 +19,17 @@ import nexusvault.format.tex.util.AwtImageConverter;
 
 public final class Tex2Png implements Converter {
 
+	private final boolean splitImages;
+	private final boolean exportMipMaps;
 	private TextureReader reader;
 	private TextureImageSplitter textureSplitter;
-	private final boolean splitImages;
 
-	public Tex2Png(boolean splitImages) {
+	public Tex2Png(boolean splitImages, boolean exportMipMaps) {
+		this.splitImages = splitImages;
+		this.exportMipMaps = exportMipMaps;
+
 		this.reader = TextureReader.buildDefault();
 		this.textureSplitter = new nexusvault.format.tex.TextureImageSplitter();
-		this.splitImages = splitImages;
 	}
 
 	@Override
@@ -38,20 +42,40 @@ public final class Tex2Png implements Converter {
 	public void convert(ConversionManager manager) throws IOException {
 		final var resource = manager.getResource();
 		final var imageObject = this.reader.read(resource.getData());
-		final var image = imageObject.getImage(0);
-		final var outputPath = manager.resolveOutputPath(PathUtil.replaceFileExtension(resource.getFile(), "png"));
+		final var images = new LinkedList<TextureImage>();
+		images.add(imageObject.getImage(0));
 
-		writeImage(image, outputPath);
-		manager.addCreatedFile(outputPath);
-
-		if (this.splitImages && this.textureSplitter.isSplitable(imageObject.getTextureDataType())) {
-			final var textureComponents = this.textureSplitter.split(image, imageObject.getTextureDataType());
-			for (int i = 0; i < textureComponents.size(); ++i) {
-				final var splitOutputPath = PathUtil.addFileNameSuffix(outputPath, String.format(".%d", i));
-				writeImage(textureComponents.get(i), splitOutputPath);
-				manager.addCreatedFile(splitOutputPath);
+		if (this.exportMipMaps) {
+			for (var i = 1; i < imageObject.getMipMapCount(); ++i) {
+				images.add(imageObject.getImage(i));
 			}
 		}
+
+		final var fileName = PathUtil.getFileName(resource.getFile());
+		for (var i = 0; i < images.size(); ++i) {
+			final var image = images.get(i);
+			final var outputPath = manager.resolveOutputPath(getFileName(fileName, i) + ".png");
+			writeImage(image, outputPath);
+			manager.addCreatedFile(outputPath);
+		}
+
+		if (this.splitImages && this.textureSplitter.isSplitable(imageObject.getTextureDataType())) {
+			for (var i = 0; i < images.size(); ++i) {
+				final var textureComponents = this.textureSplitter.split(images.get(i), imageObject.getTextureDataType());
+				for (var j = 0; j < textureComponents.size(); ++j) {
+					final var outputPath = manager.resolveOutputPath(getFileName(fileName, i) + String.format(".%d.png", j));
+					writeImage(textureComponents.get(j), outputPath);
+					manager.addCreatedFile(outputPath);
+				}
+			}
+		}
+	}
+
+	private String getFileName(String fileName, int mipmap) {
+		if (mipmap == 0) {
+			return fileName;
+		}
+		return fileName + String.format(".m%02d", mipmap);
 	}
 
 	private void writeImage(TextureImage image, Path path) throws IOException {